+++ include/mwtl/mwtl.h
@@ -1,10 +1,14 @@
 #pragma once
 
 #include <mwtl/application.h>
+#include <mwtl/command_controls.h>
+#include <mwtl/control_resources.h>
 #include <mwtl/controls.h>
 #include <mwtl/dpi.h>
 #include <mwtl/events.h>
+#include <mwtl/input_controls.h>
 #include <mwtl/message_pump.h>
+#include <mwtl/navigation_controls.h>
 #include <mwtl/timer.h>
 #include <mwtl/window.h>
 #include <mwtl/window_options.h>